* Reworked argument passing. Now you can specify multiple files to compile without having to re run the program, so your logs will not be overwritten for every WZ file you compile.  * Fixed Reactor.wz compiling giving ClassCastException.  * Fixed SkillCOnverter going into an infinite loop.  * Fixed MobConverter giving NumberFormatException (does Bodyguard B really need to drain 50000 mp even though players have a max of 30000?)  * Some integer properties were reduced in size to further save space.  * Added summons to mob skills.  * Added a way to add custom monster drops by means of editing DropConverter's source. One day, I may load custom scripts from a file so that you won't have to recompile the converter in order to change custom drops.

+++ KvjCompiler/src/kvjcompiler/reactor/ReactorConverter.java
@@ -46,7 +46,7 @@ public class ReactorConverter extends Converter {
 				if (event == XMLStreamReader.START_ELEMENT) {
 					open1++;
 					if (r.getLocalName().equals("imgdir") && r.getAttributeValue(0).equals("0")) {
-						e = new Event(Integer.parseInt(dirs[0]));
+						e = new Event(Byte.parseByte(dirs[0]));
 						for (open = 1; open > 0;) {
 							event = r.next();
 							if (event == XMLStreamReader.START_ELEMENT) {
